… fixes #2278, fixes #1800)

**Important**
Its an improvement over the old rule. I am not saying this is perfect but its way better.
It solves a lot of our issues. I have not changed the options of the rule so it will not be a breaking change.