@gyandeeps 2015-07-13
@puzrin We have inherited your rule and modified and enhanced it. Do let me know if you have any feedback.