@puzrin 2015-07-13
No problem. I already said, code can be used in eslint in any way if it helps, and even without mentioning my authorship. Thank you very much for improving upstream.