@puzrin 2015-07-13
https://github.com/nodeca/eslint-plugin-nodeca/blob/master/test/fixtures.yml

Make sure all those tests exist or have replacements.