@gyandeeps 2015-07-13
@nzakas @ilyavolodin I think I am all set now.
Do let me know when you guys ready to merge, at that time I will rebase just to make sure nothing is breaking inside master because we use `indent` rule inside our project linting.