@ilyavolodin 2015-07-13
@gyandeeps Can you run perf test before and after changes for comparison?