@gyandeeps 2015-07-13
**Before**:

```sh
Gyandeep@GYANDEEP ~/WebstormProjects/eslint (master)
$ npm run perf

> eslint@0.24.0 perf C:\Users\Gyandeep\WebstormProjects\eslint
> node Makefile.js perf

CPU Speed is 2394 with multiplier 7500000
Performance Run #1:  2370.239191ms
Performance Run #2:  2716.265021ms
Performance Run #3:  2544.124735ms
Performance Run #4:  2618.084562ms
Performance Run #5:  2435.953415ms
Performance budget ok:  2544.124735ms (limit: 3132.8320802005014ms)
```

**After**:

```sh
Gyandeep@GYANDEEP ~/WebstormProjects/eslint (IndentRuleFix)
$ npm run perf

> eslint@0.24.0 perf C:\Users\Gyandeep\WebstormProjects\eslint
> node Makefile.js perf

CPU Speed is 2394 with multiplier 7500000
Performance Run #1:  2381.150365ms
Performance Run #2:  2344.4879849999998ms
Performance Run #3:  2381.615653ms
Performance Run #4:  2351.355686ms
Performance Run #5:  2354.670434ms
Performance budget ok:  2354.670434ms (limit: 3132.8320802005014ms)
```