@ilyavolodin 2015-07-13
Very nice! That's pretty significant boost in performance!