@nzakas 2015-07-14
And the crowd applauds