@rhutchison 2015-08-11
@gyandeeps thanks for your work on this.  What are your thoughts on this case?

```javascript
foo(new Bar({
    debug: true,
    hostname: localhost
  },
  function() {
    //some code
    return;
  }
));
```