@gyandeeps 2015-08-11
@rhutchison Can you please log an issue sharing what you see right now with `eslint` and what do you expect?
Thanks